Created Pixel Pals using BFS
I have added the Breadth-First Search (BFS) algorithm to the codebase. BFS is a graph traversal algorithm that visits all the vertices of a graph in breadth-first order, i.e., it visits all the vertices at the same level before moving on to the next level. This approach ensures that the shortest path to all the vertices is found.

BFS is used in the code to efficiently traverse the graph and find the shortest path between two points. By implementing this algorithm, we are able to significantly reduce the time complexity of the code and make it more efficient.